충남대 김규현 Step1 유효성 검사, 예외 처리 구현 (#160)
* <chore>(project): 상품 관리 코드 가져오기

* <feat, style, refactor>(exception, record, repository): 상품 유무 예외처리 추가, record 클래스 가독성 좋게 변경, containsKey로직 수정

ExceptionResponseDto : 예외 응답 반환 클래스
ProductNotFoundException : 상품 존재 x 예외
GlobalExceptionHandler : Controller의 예외 처리를 수행하는 클래스
containsKey : 쿼리문 앞에 !연산자를 제거해서 해당 id를 가진 상품이 있을 때 true를 반환하게 함

* <feat>(controller, requestDto): 상품 이름 유효 판단 로직 추가

공백 포함 15자 이하
특수문자 허용 : (),[],+,-,&,/,_

* <feat>(exception): 카카오 문구 포함 상품 예외 처리

상품명에 카카오가 포함된 상품은, 담당MD와의 사전 협의가 필요하므로 사용하지 못하도록 막음

* <docs>(project): 변경, 적용할 내용
README.md 에 업데이트

* <refactor>(project): 피드백 내용 반영

1. 저는 관리자 페이지를 담당MD가 사용하고
RESTapi는 일반 사용자를 위한 것으로 구분하였습니다.
승인 방식은, 담당MD가 물품을 사전등록하는 것으로 하고,
담당MD의 역할을 가진 사용자만 카카오 문구를 넣을 수 있게 만듭니다
회원 기능 구현 시 두 유형의 사용자를 구분하는 로직을 추가하겠습니다

2. price, URL 유효성 검증 로직을 추가했습니다.
상품 최대 가격은, 상품명에 대해 피드백 주셨던 것처럼
도메인 정책에 포함되는 것 같아 Product 클래스에서 검증하게 했습니다

3. 상품명 정책을 Product 객체 내부에서 검사하게 했습니다. 이 때 던지는 예외가 기존 어노테이션이 던지던 예외와 달라 GlobalExceptionHandler에 메서드를 추가했습니다. 기존에 던지던 예외는
MethodArgumentNotValidException인데, 이 예외는
@Valid등의 Dto 유효성 검증에 실패할 때 Controller에서 발생시키는 예외로 역할이 구분되어 있어, Product 도메인에서 도메인 정책을 위해 발생시키는 예외는 다른 예외를 호출하게 했습니다.

4. 사용하지 않는 import문을 제거하고
자동 정렬 기능을 켰습니다.

5. 불필요한 주석을 제거했습니다.
Controller의 기능은 메서드, URL을 보고
API명세서와 대조해 충분히 확인 가능하다 생각해 지웠고,
Service와 repository는 메서드명을 보고 유추할 수 있다고 생각했습니다.

6. handleMethodArgumentNotValidException
의 구조를 추천해주신 stream기반으로 변경했습니다

7. 가독성을 위해 ExceptionResponseDto의 errorMessage필드를 String에서 List<String>으로 바꾸어 에러 요소마다 구분되게 하였습니다.

8. 테스트 코드를 추가했습니다

9. "카카오" 문자열은 Java의 String literal이고,
String Constant Pool에 저장되어 이 리터럴을 여러 번 사용할 때 동일한 객체가 재사용된다는 내용을 찾아보았습니다. 저는 피드백 주신 내용을
contains 메서드가 호출될 때마다 "카카오"문자열 객체가 생성된다는 뜻으로 받아들였는데요, 제가 찾아본 내용과 다른 것 같아 다시 질문드립니다.
https://deveric.tistory.com/123
https://www.baeldung.com/java-string-pool
https://hyunsb.tistory.com/66
https://codepumpkin.com/string-pool-java/

+++ src/main/java/gift/service/ProductService.java
@@ -4,6 +4,8 @@ import gift.dto.request.ProductRequestDto;
 import gift.dto.request.ProductUpdateRequestDto;
 import gift.dto.response.ProductResponseDto;
 import gift.entity.Product;
+import gift.exception.KakaoApproveException;
+import gift.exception.ProductNotFoundException;
 import gift.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
@@ -17,13 +19,25 @@ public class ProductService implements ProductServiceInterface {
         this.productRepository = productRespository;
     }
 
-    //상품 단 건 조회
+    public ProductResponseDto productToResponseDto(Product product) {
+        return new ProductResponseDto(product.productId(), product.name(), product.price(),
+            product.imageURL());
+    }
+
+
     public Product getProduct(long productId) {
+        if (!containsProduct(productId)) {
+            throw new ProductNotFoundException("상품을 찾을 수 없습니다");
+        }
         return productRepository.findById(productId).orElseThrow();
     }
 
-    //상품 추가
+
     public ProductResponseDto createProduct(ProductRequestDto productRequestDto) {
+        if (productRequestDto.name().contains("카카오")) {
+            throw new KakaoApproveException(
+                "\"카카오\" 문구가 들어간 상품은 담당MD와 협의 후 사용할 수 있습니다");
+        }
         Product product = new Product(
             productRequestDto.productId(),
             productRequestDto.name(),
@@ -31,12 +45,15 @@ public class ProductService implements ProductServiceInterface {
             productRequestDto.imageURL());
         productRepository.createProduct(product);
 
-        return product.toResponseDto();
+        return productToResponseDto(product);
     }
 
-    //상품 수정
+
     public ProductResponseDto updateProduct(long productId,
         ProductUpdateRequestDto productUpdateRequestDto) {
+        if (!containsProduct(productId)) {
+            throw new ProductNotFoundException("상품을 찾을 수 없습니다");
+        }
         Product product = new Product(
             productId,
             productUpdateRequestDto.name(),
@@ -44,16 +61,20 @@ public class ProductService implements ProductServiceInterface {
             productUpdateRequestDto.imageURL());
         productRepository.updateProduct(product);
 
-        return product.toResponseDto();
+
+        return productToResponseDto(product);
     }
 
-    //상품 삭제
+
     public void deleteProduct(long productId) {
+        if (!containsProduct(productId)) {
+            throw new ProductNotFoundException("상품을 찾을 수 없습니다");
+        }
         productRepository.delete(productId);
     }
 
-    //상품 유무 확인
+
     public boolean containsProduct(long productId) {
-        return productRepository.containsKey(productId);
+        return productRepository.productExists(productId);
     }
 }
